Consolidating history and delete functionality. Updated delete functionality and icons used. Some refactoring
