+++ exerciseTracker/build.gradle
@@ -58,6 +58,7 @@ dependencies {
     compile "com.android.support:design:$SUPPORT_LIBRARY_VERSION"
     compile "com.android.support:support-v4:$SUPPORT_LIBRARY_VERSION"
     compile "com.android.support:preference-v14:$SUPPORT_LIBRARY_VERSION"
+    compile "com.android.support:support-annotations:$SUPPORT_LIBRARY_VERSION"
 
     // RxAndroid and RxJava
     compile "io.reactivex.rxjava2:rxandroid:$RX_ANDROID"
