+++ exerciseTracker/src/main/java/com/fisincorporated/exercisetracker/ui/history/ActivityHistoryItemAdapter.java
@@ -2,11 +2,15 @@ package com.fisincorporated.exercisetracker.ui.history;
 
 import android.content.Context;
 import android.database.Cursor;
+import android.support.annotation.AnimRes;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.view.animation.Animation;
+import android.view.animation.AnimationUtils;
 import android.widget.ImageButton;
+import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.fisincorporated.exercisetracker.R;
@@ -14,15 +18,19 @@ import com.fisincorporated.exercisetracker.ui.utils.RecyclerViewCursorAdapter;
 
 import java.lang.ref.WeakReference;
 
+import static android.view.animation.Animation.AnimationListener;
+
 public class ActivityHistoryItemAdapter extends RecyclerViewCursorAdapter<ActivityHistoryItemAdapter.ItemHolder> {
 
     private Context context;
     private WeakReference<IHistoryListCallbacks> callbacks;
 
-    public ActivityHistoryItemAdapter(Context context, WeakReference<IHistoryListCallbacks> callbacks) {
+
+
+    public ActivityHistoryItemAdapter(Context context, WeakReference<IHistoryListCallbacks> historyListCallbacks) {
         super(null);
         this.context = context;
-        this.callbacks = callbacks;
+        this.callbacks = historyListCallbacks;
     }
 
     @Override
@@ -44,6 +52,7 @@ public class ActivityHistoryItemAdapter extends RecyclerViewCursorAdapter<Activi
         private TextView tvDate;
         private TextView tvDescription;
         private ImageButton imgbtnMap;
+        private ImageView imgvDeleteCheckmark;
         private boolean ignoreClick = false;
 
         public ItemHolder(View itemView, WeakReference<IHistoryListCallbacks> callbacks) {
@@ -55,9 +64,9 @@ public class ActivityHistoryItemAdapter extends RecyclerViewCursorAdapter<Activi
             tvDate = (TextView) itemView.findViewById(R.id.activity_history_row_date);
             tvDescription = (TextView) itemView.findViewById(R.id.activity_history_row_description);
             imgbtnMap = (ImageButton) itemView.findViewById(R.id.activity_history_row_btnShowMap);
+            imgvDeleteCheckmark = (ImageView) itemView.findViewById(R.id.activity_history_check_to_delete);
             summaryTextView.setOnLongClickListener(this);
             summaryTextView.setOnClickListener(this);
-
         }
 
         public void bindItem(final ActivityHistorySummary activityHistorySummary) {
@@ -68,37 +77,104 @@ public class ActivityHistoryItemAdapter extends RecyclerViewCursorAdapter<Activi
             tvDate.setText(activityHistorySummary.getActivityDate());
             tvDescription.setText(activityHistorySummary.getDescription());
             imgbtnMap.setTag(activityHistorySummary);
+            imgvDeleteCheckmark.setTag(activityHistorySummary);
+            if (callbacks.get().isSetToDelete(activityHistorySummary)){
+                imgbtnMap.setVisibility(View.GONE);
+                imgvDeleteCheckmark.setVisibility(View.VISIBLE);
+            } else {
+                imgbtnMap.setVisibility(View.VISIBLE);
+                imgvDeleteCheckmark.setVisibility(View.GONE);
+            }
+
         }
 
         @Override
         public void onClick(View v) {
-            if (ignoreClick) {
-                // was longclick
-                return;
+            if (!ignoreClick) {
+                switch (v.getId()) {
+                    case R.id.activity_row_summary_layout:
+                        viewCallbacks.get().displayStats((ActivityHistorySummary) v.getTag(), getAdapterPosition());
+                        break;
+                    case R.id.activity_history_row_btnShowMap:
+                        viewCallbacks.get().displayMap((ActivityHistorySummary) v.getTag());
+                        break;
+                }
             }
-            switch (v.getId()) {
-                case R.id.activity_row_summary_layout:
-                    viewCallbacks.get().displayStats((ActivityHistorySummary) v.getTag(), getAdapterPosition());
-                    break;
-                case R.id.activity_history_row_btnShowMap:
-                    viewCallbacks.get().displayMap((ActivityHistorySummary) v.getTag());
-                    break;
-            }
-            ignoreClick=false;
+            ignoreClick = false;
         }
 
         @Override
         public boolean onLongClick(View v) {
+            if (v.getTag() == null || !(v.getTag() instanceof ActivityHistorySummary) ) {
+                return false;
+            }
+            ActivityHistorySummary activityHistorySummary = (ActivityHistorySummary) v.getTag();
+
             switch (v.getId()) {
                 case R.id.activity_row_summary_layout:
-                    // TODO animate in checkmark
                     // toggle to delete/undelete activity
+                    if (!callbacks.get().isSetToDelete(activityHistorySummary)) {
+                        callbacks.get().deleteThisActivity(activityHistorySummary);
+                        crossFade(imgbtnMap, R.anim.rotate_fade_out, imgvDeleteCheckmark, R.anim.rotate_fade_in);
+                    } else {
+                        callbacks.get().removeFromDeleteList(activityHistorySummary);
+                        crossFade(imgvDeleteCheckmark, R.anim.rotate_fade_out, imgbtnMap, R.anim.rotate_fade_in);
+                    }
+                    // longclick also fires click event so bypass click event
                     ignoreClick = true;
                     break;
             }
             return false;
         }
+
+
+    }
+
+    private void crossFade(View outView, @AnimRes int outAnimation, View inView, @AnimRes int inAnimation) {
+        Animation fadeOutAnimation = AnimationUtils.loadAnimation(outView.getContext(), outAnimation);
+        fadeOutAnimation.setAnimationListener(new AnimationListener() {
+            @Override
+            public void onAnimationStart(Animation animation) {
+            }
+
+            @Override
+            public void onAnimationEnd(Animation animation) {
+                outView.setAlpha(0.0f);
+                outView.setVisibility(View.GONE);
+            }
+
+            @Override
+            public void onAnimationRepeat(Animation animation) {
+            }
+        });
+        outView.setAlpha(1.0f);
+        outView.setAnimation(fadeOutAnimation);
+
+
+        Animation fadeInAnimation = AnimationUtils.loadAnimation(inView.getContext(), inAnimation);
+        fadeInAnimation.setAnimationListener(new AnimationListener() {
+            @Override
+            public void onAnimationStart(Animation animation) {}
+
+            @Override
+            public void onAnimationEnd(Animation animation) {
+                inView.setAlpha(1.0f);
+            }
+
+            @Override
+            public void onAnimationRepeat(Animation animation) {}
+        });
+        inView.setAnimation(fadeInAnimation);
+        inView.setAlpha(0.0f);
+        inView.setVisibility(View.VISIBLE);
+
+        outView.startAnimation(fadeOutAnimation);
+        inView.startAnimation(fadeInAnimation);
+
+
     }
 
 
 }
+
+
