--- exerciseTracker/src/main/java/com/fisincorporated/exercisetracker/ui/history/IHistoryCallbacks.java
@@ -1,8 +1,0 @@
-package com.fisincorporated.exercisetracker.ui.history;
-
-
-public interface IHistoryCallbacks {
-
-    void displayStats(ActivityHistorySummary activityHistorySummary, int position);
-
-}
