--- exerciseTracker/src/main/java/com/fisincorporated/exercisetracker/ui/history/IHistoryDeleteCallbacks.java
@@ -1,8 +1,0 @@
-package com.fisincorporated.exercisetracker.ui.history;
-
-public interface IHistoryDeleteCallbacks extends IHistoryCallbacks {
-
-    void deleteThisActivity(ActivityHistorySummary activityHistorySummary);
-
-    boolean isSetToDelete(ActivityHistorySummary activityHistorySummary);
-}
