+++ exerciseTracker/src/main/java/com/fisincorporated/exercisetracker/ui/history/IHistoryListCallbacks.java
@@ -1,7 +1,15 @@
 package com.fisincorporated.exercisetracker.ui.history;
 
 
-public interface IHistoryListCallbacks extends IHistoryCallbacks {
+public interface IHistoryListCallbacks {
+
+    void displayStats(ActivityHistorySummary activityHistorySummary, int position);
 
     void displayMap(ActivityHistorySummary activityHistorySummary);
+
+    void deleteThisActivity(ActivityHistorySummary activityHistorySummary);
+
+    void removeFromDeleteList(ActivityHistorySummary activityHistorySummary);
+
+    boolean isSetToDelete(ActivityHistorySummary activityHistorySummary);
 }
