+++ exerciseTracker/src/main/java/com/fisincorporated/exercisetracker/ui/master/ExerciseMasterFragmentActivity.java
@@ -1,5 +1,6 @@
 package com.fisincorporated.exercisetracker.ui.master;
- 
+
+import android.graphics.drawable.ColorDrawable;
 import android.os.Bundle;
 import android.support.annotation.StringRes;
 import android.support.v4.app.Fragment;
@@ -12,55 +13,54 @@ public abstract class ExerciseMasterFragmentActivity extends ExerciseMasterActiv
 
     protected abstract Fragment createFragment();
 
-	protected int getLayoutResId() {
-		return R.layout.activity_masterdetail;
-	}
+    protected int getLayoutResId() {
+        return R.layout.activity_masterdetail;
+    }
 
-	public void onCreate(Bundle savedInstanceState) {
-		super.onCreate(savedInstanceState);
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
         // added for master/detail fragments as on tablet
-		//setContentView(R.layout.activity_fragment);
-		setContentView(getLayoutResId());
+        //setContentView(R.layout.activity_fragment);
+        setContentView(getLayoutResId());
 
-		FragmentManager fm = getSupportFragmentManager();
-		Fragment fragment = fm.findFragmentById(R.id.fragmentContainer);
-		if (fragment == null){
-			fragment = createFragment();
+        FragmentManager fm = getSupportFragmentManager();
+        Fragment fragment = fm.findFragmentById(R.id.fragmentContainer);
+        if (fragment == null) {
+            fragment = createFragment();
             if (fragment != null) {
                 fm.beginTransaction().add(R.id.fragmentContainer, fragment)
                         .commit();
             }
-		}
+        }
 
-		// implement this in superclass?
-		toolbar = (Toolbar) findViewById(R.id.toolbar);
-		if (toolbar != null) {
-			setSupportActionBar(toolbar);
-			getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-		}
-		actionBar = getSupportActionBar();
-		if (actionBar != null) {
-			actionBar.setDisplayHomeAsUpEnabled(true);
-		}
-		 
-	}
+        // implement this in superclass?
+        toolbar = (Toolbar) findViewById(R.id.toolbar);
+        if (toolbar != null) {
+            setSupportActionBar(toolbar);
+            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        }
+        actionBar = getSupportActionBar();
+        if (actionBar != null) {
+            actionBar.setDisplayHomeAsUpEnabled(true);
+            setActionBarBackgroundColorToDefault();
+        }
+    }
 
-	public void setActivityTitle(@StringRes int stringRes){
-		if (actionBar != null) {
-			actionBar.setTitle(stringRes);
-		}
+    public void setActionBarBackgroundColorToDefault() {
+        if (actionBar != null) {
+            actionBar.setBackgroundDrawable(new ColorDrawable(getResources().getColor(R.color.colorPrimary)));
+        }
+    }
 
-	}
+    public void setActivityTitle(@StringRes int stringRes) {
+        setActivityTitle(getString(stringRes));
+    }
 
-	public void setActivityTitle(String title){
-		if (actionBar != null) {
-			actionBar.setTitle(title);
-		}
+    public void setActivityTitle(String title) {
+        if (actionBar != null) {
+            actionBar.setTitle(title);
+        }
 
-	}
+    }
 
-	
-	
- 
-	
 }
