--- exerciseTracker/src/main/res/layout/activity_history_delete_list.xml
@@ -1,98 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-                xmlns:app="http://schemas.android.com/apk/res-auto"
-                android:id="@+id/activity_history_delete_list"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:animateLayoutChanges="true"
-                android:paddingLeft="@dimen/margin_double"
-                android:paddingRight="@dimen/margin_double">
-
-    <LinearLayout
-        android:id="@+id/activity_history_delete_columns"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_marginBottom="@dimen/margin_standard"
-        android:layout_marginTop="@dimen/margin_standard"
-        android:orientation="horizontal"
-        android:weightSum="100">
-
-        <LinearLayout
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="95"
-            android:orientation="horizontal"
-            android:weightSum="100">
-
-            <TextView
-                android:id="@+id/activity_history_delete_list_activity"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="30"
-                android:text="@string/exercise"
-                android:textAppearance="?android:attr/textAppearanceMedium"/>
-
-            <TextView
-                android:id="@+id/activity_history_delete_list_location"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_weight="70"
-                android:text="@string/location"
-                android:textAppearance="?android:attr/textAppearanceMedium"/>
-        </LinearLayout>
-
-    </LinearLayout>
-
-    <include
-        android:id="@+id/activity_history_delete_divider"
-        layout="@layout/divider_vertical"
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:layout_below="@+id/activity_history_delete_columns"/>
-
-    <android.support.v7.widget.RecyclerView
-        android:id="@+id/activity_history_recycler_view"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_above="@+id/activity_history_delete_buttonfooter"
-        android:layout_below="@+id/activity_history_delete_divider"
-        android:scrollbarAlwaysDrawVerticalTrack="true"
-        android:scrollbars="vertical"
-        app:layout_behavior="@string/appbar_scrolling_view_behavior"/>
-
-    <TextView
-        android:id="@+id/empty_activity_list"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_below="@+id/activity_history_delete_columns"
-        android:text="@string/no_activities_to_display"
-        android:visibility="gone"/>
-
-    <RelativeLayout
-        android:id="@+id/activity_history_delete_buttonfooter"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true">
-
-        <include
-            android:id="@+id/activity_history_delete_divider"
-            layout="@layout/divider_vertical"
-            android:layout_width="match_parent"
-            android:layout_height="1dp"
-            />
-
-        <Button
-            android:id="@+id/activity_history_delete_list_delete"
-            style="@style/standard_button_style"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentRight="true"
-            android:layout_marginBottom="@dimen/margin_standard"
-            android:layout_marginTop="@dimen/margin_standard"
-            android:enabled="false"
-            android:text="@string/delete"/>
-    </RelativeLayout>
-
-
-</RelativeLayout>
