--- exerciseTracker/src/main/res/layout/activity_history_delete_row.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              android:layout_width="match_parent"
-              android:layout_height="wrap_content"
-              android:orientation="vertical">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="?android:attr/selectableItemBackground"
-        android:orientation="horizontal"
-        android:weightSum="100">
-
-        <include
-            layout="@layout/activity_row_summary"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="95"
-            android:clickable="true"/>
-
-        <ImageView
-            android:id="@+id/activity_history_check_to_delete"
-            style="?android:attr/buttonStyleSmall"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="5"
-            android:background="@android:color/transparent"
-            android:src="@drawable/ic_check_circle"
-            android:text="@string/checked_for_deletion"
-            android:textAppearance="?android:attr/textAppearanceSmall"/>
-
-    </LinearLayout>
-
-    <include layout="@layout/divider_vertical"/>
-
-</LinearLayout>
