+++ exerciseTracker/src/main/res/layout/activity_row_summary.xml
@@ -3,6 +3,7 @@
               android:id="@+id/activity_row_summary_layout"
               android:layout_width="match_parent"
               android:layout_height="wrap_content"
+              android:background="?attr/selectableItemBackground"
               android:orientation="horizontal"
               android:weightSum="100">
 
@@ -31,9 +32,9 @@
     <LinearLayout
         android:layout_width="0dp"
         android:layout_height="wrap_content"
+        android:layout_marginStart="@dimen/margin_standard"
         android:layout_weight="70"
-        android:orientation="vertical"
-        android:layout_marginStart="@dimen/margin_standard">
+        android:orientation="vertical">
 
         <TextView
             android:id="@+id/activity_history_row_location"
