+++ exerciseTracker/src/main/res/menu/activities_list_options.xml
@@ -89,11 +89,4 @@
             </group>
         </menu>
     </item>
-    <item
-        android:id="@+id/activities_list_options_delete"
-        app:showAsAction="withText"
-        android:title="@string/delete"
-        android:visible="true">
-    </item>
-
 </menu>
