+++ exerciseTracker/src/main/res/menu/activities_trashcan.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+      xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item
+        android:id="@+id/activities_history_trashcan"
+        android:icon="@drawable/ic_delete"
+        android:title=""
+        android:visible="false"
+        app:showAsAction="ifRoom"
+        />
+</menu>
