+++ exerciseTracker/src/main/res/values/colors.xml
@@ -23,5 +23,7 @@
 
     <color name="black_overlay">#66000000</color>
 
+    <color name="check_circle_blue_grey">#607d8b</color>
+
 
 </resources>
